Merged PR 700: Error handling on Solr API calls
- Add error handling for all API calls
- Iterate all API calls - it might possible that these return with 400 in the Solr initialization phase

Overall, after Solr has just started, it seems that there might be couple of issues with calling APIs - 200 status code for getting the core status but not all information included, 500 error for core creation if the core already exists and couldn't retrieve the core information before, sometimes even 400 errors were noticed.
Because of handling this errors, now it's a working solution, but it's something what could be enhanced in the future.

Related work items: #4281